Separate logAddressIndex and logAddressIndexer
